Started working on JS Lambda function
It seems pretty likely that we'll have to do this in JS because Cognito has an
SDK to make interoperations with JS super easy, while it's really tricky in
Python.

I'm working on figuring out how to put it all together too. It seems like we
need to register google as our OpenID provider for the Cognito user pool and
that we need to handle authentication in our other lambda function. My guess is
we can just lookup the user by phone number the text was recieved from, and go
from there.

We'll also probably need to come up with a way to verify phone numbers. I don't
know anything about Amazon SNS, but it should be simple to text an auth code and
wait for them to enter it when they are signing up on the website.

We also need to figure out where creating new journals can happen from -- only
web, or also over SMS?

One final note do we want to make the phone number users text (767) 568-7625?
If you don't have a phone in front of you, that's the phone number SMS JOURNAL.
It seems like it would be a fitting number for the application!